+++ example/lib/main-localized.dart
@@ -51,7 +51,7 @@ class DemoApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // Only call clearSavedSettings() during testing to reset internal values.
-    Upgrader().clearSavedSettings();
+    Upgrader().clearSavedSettings(); // REMOVE this for release builds
 
     // On Android, setup the Appcast.
     // On iOS, the default behavior will be to use the App Store version of