+++ example/lib/main.dart
@@ -15,12 +15,13 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // Only call clearSavedSettings() during testing to reset internal values.
-    Upgrader().clearSavedSettings(); // Remove this for release builds
+    Upgrader().clearSavedSettings(); // REMOVE this for release builds
 
-    // On Android, setup the Appcast.
     // On iOS, the default behavior will be to use the App Store version of
     // the app, so update the Bundle Identifier in example/ios/Runner with a
     // valid identifier already in the App Store.
+
+    // On Android, setup the Appcast below.
     final appcastURL =
         'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
     final cfg = AppcastConfiguration(url: appcastURL, supportedOS: ['android']);