+++ lib/src/upgrade_base.dart
@@ -38,6 +38,11 @@ class UpgradeBase extends StatefulWidget {
   /// Return false when the default behavior should not execute.
   final BoolCallback? onUpdate;
 
+  /// Called when the user taps outside of the dialog and [canDismissDialog]
+  /// is false. Also called when the back button is pressed. Return true for
+  /// the screen to be popped. Not used by [UpgradeCard].
+  final BoolCallback? shouldPopScope;
+
   /// Provide an HTTP Client that can be replaced for mock testing.
   final http.Client? client;
 
@@ -50,7 +55,7 @@ class UpgradeBase extends StatefulWidget {
   /// Hide or show release notes (default: true)
   final bool? showReleaseNotes;
 
-  /// Can alert dialog be dismissed on tap outside of the alert dialog. Not used by alert card. (default: false)
+  /// Can alert dialog be dismissed on tap outside of the alert dialog. Not used by [UpgradeCard]. (default: false)
   final bool? canDismissDialog;
 
   /// The country code that will override the system locale. Optional. Used only for iOS.
@@ -74,6 +79,7 @@ class UpgradeBase extends StatefulWidget {
     this.onIgnore,
     this.onLater,
     this.onUpdate,
+    this.shouldPopScope,
     this.client,
     this.showIgnore,
     this.showLater,
@@ -112,6 +118,9 @@ class UpgradeBase extends StatefulWidget {
     }
     if (onUpdate != null) {
       Upgrader().onUpdate = onUpdate;
+    }
+    if (shouldPopScope != null) {
+      Upgrader().shouldPopScope = shouldPopScope;
     }
     if (showIgnore != null) {
       Upgrader().showIgnore = showIgnore!;